+++ src/client.ts
@@ -1,3 +1,4 @@
 "use client"
 
 export { clientify } from "./setup.client"
+export { StateVocabClientProvider } from "./provider.client"